Avoid a potential compiler warning from passing a size_t value to a printf-style function.
Patch by Benoit Hudson <bh@techhouse.brown.edu> in PR#10337.

[[originally from CVS; cvs2svn created svn r8400]]

+++ utility/astring.c
@@ -114,7 +114,8 @@ static void vadd(struct astring *astr, const char *format, va_list ap)
   char buf[1024];
 
   if (my_vsnprintf(buf, sizeof(buf), format, ap) == -1) {
-    die("Formatted string bigger than %d bytes", sizeof(buf));
+    die("Formatted string bigger than %lu bytes",
+        (unsigned long)sizeof(buf));
   }
 
   /* Avoid calling strlen with NULL. */
